+++ CHANGELOG.md
@@ -1,4 +1,11 @@
-# 0.0.3+1
+# 0.0.4
+
+- Add `session` getter to `Context` object.
+- Change checking route match to specific request (non-breaking fix).
+- Make all methods of `Context` object asynchronous.
+- Edit README.
+
+## 0.0.3+1
 
 - Edit README.
 - Rename `files` field of `Context` to `staticFiles`.
@@ -12,7 +19,7 @@
 ## 0.0.2
 
 - Change `bodyParser` to parse body in `put`, `post`, `patch` and `delete` http methods.
-- Add default error responce 501 if route for request isn't provided.
+- Add default error response 501 if route for request isn't provided.
 
 ## 0.0.1